docs: update (#27)
* docs: update

This commit fixes spelling on the bootloader unlock section.

* docs: update

This commit fixes spelling on the installation instructions section.

* docs: update

This commit fixes spelling on the installation instructions section.

* docs: update

This commit fixes spelling on the bootloader unlock section.

* docs: update

This commit fixes spelling on the bootloader unlock section.

* docs: update

This commit fixes spelling in the Installation (by Download mode) section.