docs: update (#35)
* docs: update

This commit fixes spelling on the bootloader unlock section.

* docs: update

This commit fixes spelling on the installation instructions section.

* docs: update

This commit fixes spelling on the installation instructions section.

* docs: update

This commit fixes spelling on the bootloader unlock section.

* docs: update

This commit fixes spelling on the bootloader unlock section.

* docs: update

This commit fixes spelling in the Installation (by Download mode) section.

* docs: update

This commit fixes spelling in the Install UN1CA (Custom recovery) section.

* docs: update

This commit fixes spelling in the Install UN1CA (Download mode) section.

* docs: update

* docs: update