platforms: exynos{9820,990}: Build kernel in-tree
ExtremeROM and ExtremeKernel need to be tested hand-in-hand, as not
doing this creates feature imparity.

Build ExtremeKernel in-tree. Use a new kernel_tmp dir which should
not be cleaned unless you are changing SoCs. Cleanbuild/changes are
handled by the Kernel's Kbuild system. As such, this change adds
~3-5 minutes extra build time on the first build, but only 10-30s on
all subsequent builds.

Signed-off-by: Igor Belwon <igor.belwon@mentallysanemainliners.org>