platform: exynos2100: Build kernel in-tree
ExtremeROM and ChicletKernel need to be tested hand-in-hand, as not
doing this creates feature imparity.

Build ChicletKernel in-tree. Use a new kernel_tmp dir which should
not be cleaned unless you are changing SoCs. Cleanbuild/changes are
handled by the Kernel's Kbuild system. As such, this change adds
~3-5 minutes extra build time on the first build, but only 10-30s on
all subsequent builds.

Signed-off-by: Fred <74693315+xfwdrev@users.noreply.github.com>